+++ src/week09/onchain/oracle.py
@@ -1,4 +1,4 @@
-from src.week09.on_chain.common import *
+from src.week09.onchain.common import *
 
 
 @dataclass()
